[skip ci] Moved query parameter exporter tests
no ref

Separating subprocess coverage from policy validation keeps the schema tests focused and exercises the public pnpm export command.
